requirements.txt
Django==3.2.5
django_redis==5.0.0
pyrebase==3.0.27
python-dotenv==0.18.0
whitenoise==5.3.0
gunicorn==20.1.0
django-heroku==0.3.1
celery==5.1.2
